% Auto-generated: do not edit by hand
\docType{package}
\name{dashBio-package}
\alias{dashBio}
\title{Open-source Bioinformatics Components for Dash
}
\description{
Dash Bio is a suite of bioinformatics components that make it simpler to analyze and visualize bioinformatics data and interact with them in a Dash application.

}
\author{
\strong{Maintainer}: Ryan Patrick Kyle <ryan@plot.ly>
}
