% Auto-generated: do not edit by hand
\name{dashbioJsme}

\alias{dashbioJsme}

\title{Jsme component}

\description{
JSME is a molecule editor that supports drawing and editing of molecules and reactions on desktop computer, as well as on handheld devices including iPhone, iPad and Android smartphones and tablets. A built-in substituent menu and several keyboard shortcuts provide speedy access to the most common editing features and allow easy and fast creation of even large and complex molecules. The editor is able to export molecules as SMILES, MDL/Symyx/Accelrys Molfile or in its own compact format (one line textual representation of a molecule or reaction including also atomic 2D coordinates). The SMILES code generated by the JSME is canonical, i.e. independent on the way how the molecule was drawn.  See more detailed documentation here: https://jsme-editor.github.io/help.html
}

\usage{
dashbioJsme(id=NULL, eventSmiles=NULL, height=NULL, options=NULL,
smiles=NULL, style=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components in callbacks.
The ID needs to be unique across all of the components in an app.}

\item{eventSmiles}{Character. A Dash prop that returns data when SMILE will be changed.}

\item{height}{Character. The height of the JSME container. Can be set in px, \% etc.}

\item{options}{Character. String that is a comma separated string of JSME options.
The available options are described on the
https://wiki.jmol.org/index.php/Jmol_JavaScript_Object/JME/Options}

\item{smiles}{Character. The molecule SMILE to display}

\item{style}{Named list. Generic style overrides on the plot div}

\item{width}{Character. The width of the JSME container. Can be set in px, \% etc.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

