% Auto-generated: do not edit by hand
\name{dashbioNeedlePlot}

\alias{dashbioNeedlePlot}

\title{NeedlePlot component}

\description{

}

\usage{
dashbioNeedlePlot(id=NULL, mutationData=NULL, xlabel=NULL, ylabel=NULL, rangeSlider=NULL, needleStyle=NULL, domainStyle=NULL)
}

\arguments{
\item{id}{The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{mutationData}{}

\item{xlabel}{}

\item{ylabel}{}

\item{rangeSlider}{}

\item{needleStyle}{}

\item{domainStyle}{}
}
