% Auto-generated: do not edit by hand
\docType{package}
\name{dashBio-package}
\alias{dashBio}
\title{Open-source Bioinformatics Components for Dash}
\description{
A suite of components that makes it simpler to analyze and visualize bioinformatics data and interact with them in a Dash application.
}
\author{
\strong{Maintainer}: Ryan Patrick Kyle <ryan@plot.ly>
}
