% Auto-generated: do not edit by hand
\name{dashbioNeedlePlot}

\alias{dashbioNeedlePlot}

\title{NeedlePlot component}

\description{
The Needle Plot component is used to visualize large datasets containing categorical or numerical data. The lines and markers in the plot correspond to bars in a histogram.
}

\usage{
dashbioNeedlePlot(id=NULL, mutationData=NULL, xlabel=NULL, ylabel=NULL, rangeSlider=NULL, needleStyle=NULL, domainStyle=NULL)
}

\arguments{
\item{id}{The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{mutationData}{The data that are displayed on the plot}

\item{xlabel}{Title of the x-axis.}

\item{ylabel}{Title of the y-axis.}

\item{rangeSlider}{If true, enables a rangeslider for the x-axis.}

\item{needleStyle}{Options for the needle marking single site mutations}

\item{domainStyle}{Options for the protein domain coloring}
}

